axi_logic_analyzer: Tune trigger delay
The trigger signal that goes to the DMA(fifo_wr_sync) does not pass through
the variable fifo, for this reason, a 3 clock cycles delay is required, to
keep in sync the data with the trigger.
On the other hand, to be able to cascade the axi_logic_analyzer with
axi_adc_trigger, there should not be any delay on the trigger path, for this
purpose, the trigger_out_express was created.

Remove the extra delays on the trigger_i.